Alertify added for product delete and css changes :saxophone:
Alertify added for product delete and css changes :saxophone:

+++ script/js/actions/ContentAction.js
@@ -1,5 +1,10 @@
 var deleteProductButtonClicked = function (sProductId) {
-  ContentListStore.deleteProductById(sProductId);
+  alertify.confirm("Are you sure you want to delete this product?",function(eventDelete){
+    if(eventDelete) {
+      ContentListStore.deleteProductById(sProductId);
+      alertify.success("Product successfully deleted");
+    }
+  });
 };
 
 var handleViewButtonClicked = function(viewStyleButtonEvent) {
